feat: Add movie detail activity, make MovieResult parcelable
MovieDetail was changed to implement parcelable so that a reference to an object (and its properties) can be passed through an intent, to allow for the object to be referenced from within the new activity. Also setup a basic MovieDetailResult activity that loads basic movie details which it is launched when a posted from the main pages is clicked / selected.